Restructure folder layout to improve readability
